refactor(api): move disable certification center membership usecase to team context
